Add additional exception handling to avoid serverside issues. (#1579)
* Add additional exception handling to avoid serverside issues.


* Run cleanup on all projects

+++ analytics/api/src/main/java/org/locationtech/geowave/analytic/sample/BahmanEtAlSampleProbabilityFn.java
@@ -22,7 +22,7 @@ public class BahmanEtAlSampleProbabilityFn implements SampleProbabilityFn {
       final double weight,
       final double normalizingConstant,
       final int sampleSize) {
-    return (((double) sampleSize) * weight) / normalizingConstant;
+    return ((sampleSize) * weight) / normalizingConstant;
   }
 
   @Override
